refactor: final structure
