+++ batcher/.gitignore
@@ -1,3 +1,2 @@
-.env
+aligned-batcher/.env
 */node_modules
-
