+++ batcher/Cargo.toml
@@ -1,40 +1,7 @@
-[package]
-name = "batcher"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "batcher"
-path = "src/lib.rs"
-
-[[bin]]
-name = "batcher"
-path = "src/main.rs"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio-tungstenite = "0.21.0"
-futures-channel = "0.3.30"
-futures-util = "0.3.30"
-tokio = { version = "1.37.0", features = ["rt", "rt-multi-thread", "macros"] }
-log = "0.4.21"
-env_logger = "0.11.3"
-serde_json = "1.0.117"
-serde = { version = "1.0.201", features = ["derive"] }
-clap = { version = "4.5.4", features = ["derive"] }
-sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", rev = "v1.0.0-testnet" }
-bincode = "1.3.3"
-sha3 = "0.10.8"
-aws-config = "1.4.0"
-aws-sdk-s3 = "1.29.0"
-bytes = "1.6.0"
-hex = "0.4.3"
-dotenv = "0.15.0"
-anyhow = "1.0.83"
-ethers = { version = "2.0", features = ["ws", "rustls"] }
-lambdaworks-crypto = { version = "0.7.0", features = ["serde"] }
-rand = "0.8.5"
-serde_yaml = "0.9.34"
-alloy-primitives = "0.7.4"
-lazy_static = "1.4.0"
+[workspace]
+members = [
+    "aligned-batcher",
+    "aligned-batcher-client",
+    "aligned-batcher-lib"
+]
+resolver = "2"
