+++ batcher/aligned-batcher-client/Cargo.toml
@@ -12,10 +12,10 @@ tokio-tungstenite = "0.21.0"
 tungstenite = "0.21.0"
 url = "2.5.0"
 futures-channel = "0.3.30"
-batcher = { path = "../" }
 bytes = "1.6.0"
 log = "0.4.21"
 env_logger = "0.11.3"
 anyhow = "1.0.83"
 alloy-primitives = "0.7.4"
 clap = { version = "4.5.4", features = ["derive"] }
+aligned-batcher-lib = { path = "../aligned-batcher-lib" }
