+++ batcher/aligned-batcher-client/send_burst_tasks.sh
@@ -20,4 +20,3 @@ do
   sleep 1
   counter=$((counter + 1))
 done
-
