+++ batcher/aligned-batcher-lib/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "aligned-batcher-lib"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+alloy-primitives = "0.7.4"
+anyhow = "1.0.83"
+lazy_static = "1.4.0"
+log = "0.4.21"
+serde = { version = "1.0.201", features = ["derive"] }
+sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", rev = "v1.0.0-testnet" }
+bincode = "1.3.3"
+hex = "0.4.3"
