+++ batcher/aligned-batcher/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "aligned-batcher"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+tokio-tungstenite = "0.21.0"
+futures-util = "0.3.30"
+tokio = { version = "1.37.0", features = ["rt", "rt-multi-thread", "macros"] }
+log = "0.4.21"
+env_logger = "0.11.3"
+serde_json = "1.0.117"
+serde = { version = "1.0.201", features = ["derive"] }
+clap = { version = "4.5.4", features = ["derive"] }
+sha3 = "0.10.8"
+aws-config = "1.4.0"
+aws-sdk-s3 = "1.29.0"
+bytes = "1.6.0"
+hex = "0.4.3"
+dotenv = "0.15.0"
+anyhow = "1.0.83"
+ethers = { version = "2.0", features = ["ws", "rustls"] }
+lambdaworks-crypto = { version = "0.7.0", features = ["serde"] }
+serde_yaml = "0.9.34"
+aligned-batcher-lib = { path = "../aligned-batcher-lib" }
