+++ batcher/aligned-batcher/gnark/verifier.go
